# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Brand Service API",
    "description": "Microservice for managing brand information, media assets, organization data, and AI-powered content analysis.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "schemas": {
...
      "GetBrandResponse": {
        "type": "object",
        "properties": {
          "brand": {
            "$ref": "#/components/schemas/BrandDetail"
          }
        },
        "required": [
          "brand"
        ]
      },
      "BrandDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Brand UUID"
          },
          "domain": {
            "type": "string",
            "nullable": true,
            "description": "Normalized domain (subdomains preserved, www stripped). `null` for a no-website brand (identified by `name`)."
          },
          "url": {
            "type": "string",
            "nullable": true,
            "description": "Full brand website URL. `null` for a no-website brand."
          },
          "name": {
            "type": "string",
            "description": "Brand display name. Lazy-extracted from the website on first read if missing; user-provided for a no-website brand."
          },
          "logoUrl": {
            "type": "string",
            "nullable": true,
            "description": "Logo image URL. Lazy-filled with a deterministic logo.dev URL on first read if missing. `null` for a no-website brand (no domain to build one from)."
          },
          "clickDestinationUrl": {
            "type": "string",
            "nullable": true,
            "description": "Page outreach clicks should land on. Defaults to the brand's own landing URL (`url`) when the user has not set an override. `null` only for a no-website brand with no override (no landing URL to fall back to). Per-brand config, set via PUT /orgs/brands/{brandId}/click-destination."
          },
          "whatsAppLink": {
            "type": "string",
            "nullable": true,
            "description": "The brand's WhatsApp link — the click destination for the \"maximize WhatsApp conversations\" goal. `null` when unset (no sensible default, unlike clickDestinationUrl). Per-brand config, set via PUT /orgs/brands/{brandId}/whatsapp-link."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp when the brand row was created."
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO timestamp when the brand row was last updated."
          }
        },
        "required": [
          "id",
          "domain",
          "url",
          "name",
          "logoUrl",
          "clickDestinationUrl",
          "whatsAppLink",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    }
  },
  "paths": {
...
    "/public/brands/{id}": {
      "get": {
        "summary": "Get a single brand by ID (public, no auth)",
        "description": "Public mirror of GET /internal/brands/{id}. Identical response shape — identity + lazy-filled name and logoUrl. Use this when no API key is available (dashboards, embeddable widgets). Business fields must still be fetched via POST /orgs/brands/extract-fields (org auth required).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "orgId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBrandResponse"
                }
              }
            }
          },
          "404": {
            "description": "Brand not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}